backend/requirements.txt
Django==5.0.7
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.5.0
PyMySQL==1.1.0
python-dotenv==1.0.1
